src/tg/requirements.txt
# Telebot
aiogram

# Security
python-dotenv

# Linter and Formatter
ruff